20-classes/requirements.txt
anyio==4.9.0
attrs==25.3.0
black==25.1.0
click==8.2.1
idna==3.10
mypy==1.17.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pyright==1.1.403
ruff==0.12.5
sniffio==1.3.1
typing_extensions==4.14.1
watchfiles==1.1.0
